# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "project stor api ali_naseri_php",
        "version": "0.0.0"
    },
    "paths": {
...
        "/api/v1/proppertis": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "proppertis all ",
                "description": "proppertis all",
                "operationId": "allproppertis",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "all proppertis"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "proppertis store ",
                "description": "proppertis store",
                "operationId": "storeproppertis",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "categorie_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "store proppertis"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/proppertis/2": {
            "put": {
                "tags": [
                    "categories"
                ],
                "summary": "proppertis updated Number 2 can be put the other thing",
                "description": "proppertis updated",
                "operationId": "updateproppertis",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "categorie_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "all proppertis"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/images": {
            "get": {
                "tags": [
                    "images"
                ],
                "summary": "images all",
                "description": "images all",
                "operationId": "imagesall",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "all url amage"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "post": {
                "tags": [
                    "images"
                ],
                "summary": "images store",
                "description": "images store",
                "operationId": "imagesstore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "type": "image"
                                    },
                                    "imageable_type": {
                                        "type": "string"
                                    },
                                    "imageable_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "save ok !"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/images/2": {
            "put": {
                "tags": [
                    "images"
                ],
                "summary": "images update !",
                "description": "images update",
                "operationId": "imagesupdate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "type": "image"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "save update ok !"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        }
    }
}